- fixed data generation, minor problems remain in trelis, will be fixed in resonable time

+++ analysis/trelis.py
@@ -66,7 +66,7 @@ class Trelis(object):
             index = np.where(states == np.min(states))[0]
             output[data_point_index] = np.mod(index[0],constellation_size) + 1
 
-        return output
+        return output.astype(int)
